requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
lxml==3.8.0
olefile==0.44
Pillow==4.2.1
simplejson==3.11.1
